+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## 0.2.8 - 13-Jun-2025
 * Minor optimization where I avoid making repeated calls.
+* Add a keys method delegator.
 
 ## 0.2.7 - 20-Apr-2024
 * Rubocop updates.